app/requirements.txt
flask==2.3.2
Flask-Cors==6.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2  # Must upgrade from 0.2.0
google-auth==2.22.0
google-api-python-client==2.104.0  # Required dependency
